Fixed webView callback bug, improved rendering
After each login _onNavigationStateChange is called three times.
First time when webview url is changed, then after loading started
and at last when loading is done.
This resulting in three call to the instagram API with one success
and two errors.

The rendering was also changed to always render the webview but with
a 0% height at first. That way the webview and instagram login page
is loaded before our user click on the main page login button,
resulting in a hidden web page loading time.

A last change was made when the login view is hidden, we change the
key state value in order to trigger a new render of the webview
for resetting url to initial instagram login